refactor: rename events.py to controller_events.py
Mechanical rename to free up event.py (singular) for the upcoming
event platform port from core. The IsyControllerEvents hub remains
the central dispatcher for node/program/system events; only the
filename and import paths change.
